tests: Move common test elements to a seperate directory.
